Started two processes:
1. Adding function composition to the code
I started to realize that there are a lot of places with nested function
calls (that is: f(g(x))),
and this can be improved via function composition.
It will also make the code a lot longer :D.
Some functions have many layers of nesting, which will require a more
complex composition structure.
I'm thinking of implementing singly linked lists and folding over them.

2. Creating an automatic translator
Believe it or not, I translated the source code to the one-liner by hand.
With the addition of complex function composition, doing this by hand is
simply not worth it any more. So, I'm building a translator.
This requires the creation of is_even_src.py, which will include
unannotated code for translation, and translator.py, which will
translate is_even_src.py into a one-liner.